msm: kgsl: Affine IRQ and worker kthread to the big CPU cluster
These are in the critical path for rendering frames to the display, so
mark them as performance-critical and affine them to the big CPU
cluster. They aren't placed onto the prime cluster because the
single-CPU prime cluster will be used to run the DRM IRQ and kthreads.
DRM is more latency-critical than KGSL and we need to have DRM and KGSL
running on separate CPUs for the best performance, so KGSL gets the big
cluster.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Change-Id: Ia01ca9283f339572a247593d0f3ab92fc236e69c
Signed-off-by: fadlyas07 <mhmmdfdlyas@proton.me>