workqueue: Affine unbound workqueues to little CPUs by default
Although unbound workqueues are eligible to run their workers on any
CPU, the workqueue subsystem prefers scheduling workers onto the CPU
which queues them. This results in unbound workers consuming valuable
CPU time on the big and prime CPU clusters.

We can alleviate the burden of kernel housekeeping on the more important
CPUs by moving the unbound workqueues to the little CPU cluster by
default. This may also reduce power consumption, which is a plus.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[0ctobot: Drop unused hk_flags function]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Change-Id: I1fd6535f6a6cd2e1feb5cbd1683afada76448533
Signed-off-by: fadlyas07 <mhmmdfdlyas@proton.me>