sched/fair: Don't remove important task migration logic from PELT
This task migration logic is guarded by a WALT #ifdef even though it has
nothing specific to do with WALT. The result is that, with PELT, boosted
tasks can be migrated to the little cluster, causing visible stutters.

Move the WALT #ifdef so PELT can benefit from this logic as well.

Thanks to Zachariah Kennedy <zkennedy87@gmail.com> and Danny Lin
<danny@kdrag0n.dev> for discovering this issue and creating this fix.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Change-Id: I0cc4ccbb14ac2d23267a72b31b99db30525973d7
Signed-off-by: fadlyas07 <mhmmdfdlyas@proton.me>