sched/core: Skip rq lock in try_to_wake_up() when WALT is disabled
When WALT is disabled, do_pl_notif() always returns false, in which case
this bit of code serves no purpose. As a result, this WALT-specific code
spins on acquiring the rq lock in a hot path, wasting CPU time. Compile
it out when WALT is disabled to eliminate the unnecessary overhead.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Change-Id: I408b7fabf96d1dc5fad71fbe9a08434b6481f3c3
Signed-off-by: fadlyas07 <mhmmdfdlyas@proton.me>